backend/requirement_tutoring.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
decorator==4.4.1 
Django==2.2.6
django-multiselectfield==0.1.10
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.4
Pillow==6.2.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
travis==0.0.3
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.7
uWSGI==2.0.18
wrapt==1.11.2
